ctr: move tasks, run to commands package
